refactor: refactored save box dialog
