+++ app/playground/BoxPage.tsx
@@ -1,21 +1,17 @@
 'use client';
 
-import { ComponentTag } from '@algo-sandbox/core';
 import AppNavBar from '@components/AppNavBar';
 import { BoxExecutionControls, useBoxContext } from '@components/box-page';
+import { SaveBoxDialog } from '@components/box-page/SaveBoxDialog';
 import SettingsDialog from '@components/common/SettingsDialog';
 import { useSandboxComponents } from '@components/playground/SandboxComponentsProvider';
 import { useTabManager } from '@components/tab-manager/TabManager';
 import TabProvider from '@components/tab-manager/TabProvider';
-import { Button, Input, TagInput } from '@components/ui';
-import Dialog from '@components/ui/Dialog';
 import { TabsItem } from '@components/ui/VerticalTabs';
-import getSandboxObjectConfig from '@utils/getSandboxObjectConfig';
 import groupOptionsByTag from '@utils/groupOptionsByTag';
 import clsx from 'clsx';
 import { useRouter, useSearchParams } from 'next/navigation';
-import { useEffect, useMemo, useState } from 'react';
-import { Controller, useForm } from 'react-hook-form';
+import { useMemo, useState } from 'react';
 import { toast } from 'sonner';
 
 import { BoxPageExecutionWrapper } from './BoxPageExecutionWrapper';
@@ -26,12 +22,7 @@ import { BoxPageSidebar } from './BoxPageSidebar';
 function BoxPageImpl() {
   const boxKey = useSearchParams().get('box') ?? '';
   const router = useRouter();
-  const {
-    isBoxCustom,
-    save: saveBox,
-    isBoxDirty,
-    delete: deleteBox,
-  } = useBoxContext();
+  const { isBoxCustom, delete: deleteBox } = useBoxContext();
   const { boxOptions } = useSandboxComponents();
   const groupedBoxOptions = useMemo(() => {
     return groupOptionsByTag(boxOptions, { omitTags: ['box'] });
@@ -44,39 +35,7 @@ function BoxPageImpl() {
   const hasBox = selectedOption !== undefined;
   const [showSettings, setShowSettings] = useState(false);
 
-  const selectedOptionTags = useMemo(() => {
-    if (selectedOption === undefined) {
-      return [];
-    }
-
-    return getSandboxObjectConfig(selectedOption.value).tags;
-  }, [selectedOption]);
-
   const [saveBoxDialogOpen, setSaveBoxDialogOpen] = useState(false);
-
-  const {
-    register,
-    handleSubmit,
-    control,
-    getValues,
-    reset,
-    formState: { isDirty },
-  } = useForm<{
-    name: string;
-    tags: Array<ComponentTag>;
-  }>({
-    defaultValues: {
-      name: selectedOption?.label ?? '',
-      tags: selectedOptionTags,
-    },
-  });
-
-  useEffect(() => {
-    reset({
-      name: selectedOption?.label ?? '',
-      tags: selectedOptionTags,
-    });
-  }, [selectedOption, reset, selectedOptionTags]);
 
   const {
     selectedTabId,
@@ -113,16 +72,6 @@ function BoxPageImpl() {
 
   const handleSaveClick = () => {
     setSaveBoxDialogOpen(true);
-  };
-
-  const handleSaveBox: typeof saveBox = async (options) => {
-    setSaveBoxDialogOpen(false);
-    await saveBox(options);
-    if (options.asNew) {
-      toast.success(`Saved as "${options.name}"`);
-    } else {
-      toast.success(`Saved changes to ${options.name}`);
-    }
   };
 
   const handleDeleteClick = async () => {
@@ -182,75 +131,24 @@ function BoxPageImpl() {
             </TabProvider>
           ))}
         </div>
+        {/* Controls for larger screens */}
         <div className="fixed bottom-4 left-1/2 -translate-x-1/2 z-50">
           <div className="hidden lg:flex items-center gap-4 rounded-full px-4 py-2 bg-surface border shadow">
             <BoxExecutionControls />
           </div>
         </div>
+        {/* Controls for small screens */}
         {isExecutionPageVisible && (
           <div className="flex px-2 gap-4 justify-center border-t items-center py-2 lg:hidden">
             <BoxExecutionControls />
           </div>
         )}
       </div>
-      <Dialog
-        title="Save box"
-        content={
-          <form
-            onSubmit={handleSubmit((data) => {
-              handleSaveBox({ ...data, asNew: !isBoxCustom });
-            })}
-          >
-            <Input
-              label="Name"
-              containerClassName="flex-1 mb-2"
-              {...register('name')}
-            />
-            <Controller
-              control={control}
-              name="tags"
-              render={({ field }) => (
-                <TagInput
-                  label="Tags (enter to add)"
-                  value={field.value}
-                  onChange={(tags) => {
-                    field.onChange({
-                      target: {
-                        value: tags,
-                      },
-                    });
-                  }}
-                />
-              )}
-            />
-            <div className="flex gap-2 mt-4 justify-end">
-              {!isBoxCustom && (
-                <Button type="submit" variant="primary" label="Save as new" />
-              )}
-              {isBoxCustom && (
-                <>
-                  <Button
-                    type="button"
-                    variant="filled"
-                    label="Save as new"
-                    onClick={() => {
-                      handleSaveBox({ ...getValues(), asNew: true });
-                    }}
-                  />
-                  <Button
-                    label="Save"
-                    variant="primary"
-                    type="submit"
-                    disabled={!isBoxDirty && !isDirty}
-                    onClick={handleSaveClick}
-                  />
-                </>
-              )}
-            </div>
-          </form>
-        }
-        open={saveBoxDialogOpen}
-        onOpenChange={setSaveBoxDialogOpen}
+      <SaveBoxDialog
+        selectedOption={selectedOption}
+        saveBoxDialogOpen={saveBoxDialogOpen}
+        setSaveBoxDialogOpen={setSaveBoxDialogOpen}
+        handleSaveClick={handleSaveClick}
       />
     </>
   );
